Finally figured out how to separate capture templates
I misunderstood how ` worked.  I knew it marked something as a list,
and I knew that it somehow allowed you to execute things inside that
list, but I didn't understand how it interplayed with ,.  Inside a
list that has been declared with a backtick (`), anything that follows
a comma (,) is considered "executed". Sort of the opposite of the
original quote (').

In this case, Org Mode expects the contents of org-capture-templates
to be a list containing elements that are in a certain format, NOT a
list containing variables that have that format. For whatever reason,
we need things to be inline. We can accomplish that by declaring the
list with the backtick (`) syntax and then preceeding each element in
the list with a commna (,) which "evaluates" the variables to their
values and all is right with the world. I can now separate each of
these templates out into its own src block in org mode and document
them individually!