Added unit tests to my config (YAY)
I didn't realize there was a built in test framework for elisp!! This
is amazing! Except, the first test I wanted to write I needed to mock
a function and I don't know of a good way to "mock" that function
that won't break the setup after the function runs, so I just skipped
the test on anything that isn't linux for now.