Defer loading spacemacs-theme
It seems like use-package was trying to activate the plugin
immediately, which was causing an exception to be thrown. defering it
seems to allow something else (possibly the built in theme loading
functions?) to be the thing that actually activates the plugin. I'm
unusure, though, the actual order of how this is being executed now.