Fixing fonts
Fonts weren't being set for new frames as well as the existing
frame. This fixes that. I don't totally understand why because the
documentation makes me think that this should only impact frames that
exist at the time the function runs, but it seems to be having the
desired effect.