Added visual fill mode
On mobile, files that I wrote in Markdown were getting annoying to
read. The line breaks at 80 columns would just break in random places
on a screen of a very different width and that was annoying. these
fill modes cause Emacs to soft-wrap the text visually when  I am
writing in Emacs without actually adding a newline, so that my stuff
on mobile doesn't look awful.