Fixed drawer config
It was bothering me that my notes (C-c C-z in edit mode or a in agenda
mode) were being tossed under the logbook drawer. I have turned off
logging repeats in the logbook, which is part of why I had set this in
the first place. Now that I'm not tracking the repeats, having notes
added under logbook is annoying, so I am undoing it.